% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.R
\name{Distance}
\alias{Distance}
\title{Distance}
\usage{
Distance(x, y, z, k, l, m)
}
\arguments{
\item{x}{a value on axis 1, pairs with \code(k)}

\item{y}{a value on axis 2, pairs with \code(l)}

\item{z}{a value on axis 3, pairs with \code(m)}

\item{k}{a value on axis 2, pairs with \code(y)}

\item{l}{a value on axis 1, pairs with \code(x)}

\item{m}{a value on axis 3, pairs with \code(z)}
}
\description{
calculation of geometric distance for three pairs of values describing two points
}
\author{
Greg Pilgrim \email{gpilgrim@vergason.com}
}
